refactor(claw-server-rust): flatten database module
